frontend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
psycopg2-binary==2.9.9